Introduce check_multiple_package_version actor
Upgrade should be inhibited if both x32 and x64 version of certain packages
are simultaneously present on the machine.
The actor will check for those packages and produce a Report with a bugzilla
link (if any is present for the problematic package).
The implementation also enhances the create_lookup library function allowing it
to create multi-parameter lookups.

TBD:
- create_lookup tests in test_redhatsignedrpmscanner.py appear to be completely
meaningless and need to be refactored & updated.
- check_multiple_package_version tests

OAMG-4097