storagescanner: Inhibit if any CIFS entries found in /etc/fstab
Due to networking not available in initramfs phase CIFS upgrade attempt
with CIFS entries in the /etc/fstab has to be inhibited for now.
Also changed a behavior of error reporting in _get_fstab_info
function, instead of find-first-error-and-report-it now all errors
of the same type will be reported under a single report entry.

OAMG-4107